Generated by Spark: i need to make a web app that or just an HTML page that mirrors sections of two other pages so that i can see both weather information areas at the same time:     http://161.99.100.251/admin/mainreadouts.php that is EOA Weather http://161.99.100.30/admin/mainreadouts.php that is WOA Weather
the image i provided was the area on the two intraweb sights that i want to mirror in the new one. i want just one place for both weather stations and not have to go to two websites. also when you access the sites there is a simple user name and password, they area username: admin password: admin